为idle_chat增加manager，负责管理实例
