+++ src/experimental/PFC/conversation_initializer.py
@@ -18,6 +18,7 @@ from .observation_info import ObservationInfo
 from .conversation_info import ConversationInfo
 from .reply_generator import ReplyGenerator
 from .PFC_idle.idle_chat import IdleChat
+from .PFC_idle.idle_manager import IdleManager
 from .waiter import Waiter
 from .pfc_utils import get_person_id
 from .reply_checker import ReplyChecker
@@ -189,8 +190,12 @@ async def initialize_core_components(conversation_instance: "Conversation"):
         conversation_instance.idle_chat = IdleChat.get_instance(
             conversation_instance.stream_id, conversation_instance.private_name
         )
-        await conversation_instance.idle_chat.increment_active_instances()
-        logger.debug(f"[私聊][{conversation_instance.private_name}] (Initializer) IdleChat实例已获取并增加活跃计数")
+        await IdleManager._global_lock.acquire()
+        try:
+            IdleManager._global_active_instances_count += 1
+            logger.debug(f"[私聊][{conversation_instance.private_name}] (Initializer) IdleChat实例已获取并增加活跃计数，当前计数：{IdleManager._global_active_instances_count}")
+        finally:
+            IdleManager._global_lock.release()
 
         # 2. 初始化信息存储和观察组件
         logger.debug(f"[私聊][{conversation_instance.private_name}] (Initializer) 获取 ChatObserver 实例...")
